# go/swag-go-example/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple swagger implementation in Go HTTP",
        "title": "Go Restful API with Swagger",
        "contact": {
            "name": "Linggar Primahastoko",
            "url": "http://linggar.asia",
            "email": "x@linggar.asia"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Auth Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Auth Login",
                "operationId": "auth-login",
                "parameters": [
                    {
                        "description": "Auth Login Input",
                        "name": "authLogin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.authParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.authResp"
                        }
                    }
                }
            }
        },
        "/hello": {
            "get": {
                "description": "hello test",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "hello handler",
                "operationId": "hello-world",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "main.authParam": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "main.authResp": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}